Spiking improved resource abstraction - simplify directory listing
